% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemlm.R
\name{chemlm}
\alias{chemlm}
\alias{chemlm.default}
\alias{print.chemlm}
\alias{plot.chemlm}
\title{chemlm}
\usage{
chemlm(x, ...)

\method{chemlm}{default}(
  data,
  outcome,
  chem,
  value = "value",
  adjust = NULL,
  confound = NULL,
  family = "gaussian",
  type = "filter"
)

\method{print}{chemlm}(x)

\method{plot}{chemlm}(x, scales = "free", ncol = 3, facetchem = F)
}
\arguments{
\item{data}{dataset.  data should be in "long" format with the chemical names in one column and the values in another}

\item{outcome}{character outcome variable}

\item{chem}{column name for chemical type (string format)}

\item{value}{column name for value (string format)}

\item{adjust}{function to modify chemicals (default is NULL)}

\item{confound}{character vector of confound variables}

\item{family}{regression family (default is linear model)}
}
\description{
Regression models for multiple chemical exposures
}
\details{
\code{chemlm} Run regression model for chemical data

This is a function to run multiple regression models for chemical data,
where each model is fitted to a single chemical at a time
}
\examples{
data(simchemdat)
res <- chemlm(simchemdat, outcome = "out", chem = "chem")
print(res)
plot(res)
}
