% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemregr-package.R
\docType{package}
\name{chemregr}
\alias{chemregr}
\alias{chemregr-package}
\title{Regression models for multiple chemical exposures}
\description{
This package contains functions to facilitate estimating associations
between multiple chemical exposures and a health outcome
}
