% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemlm.R
\name{innerchem}
\alias{innerchem}
\title{innerchem}
\usage{
innerchem(
  data,
  outcome,
  value = "value",
  confound = NULL,
  family = "gaussian",
  type = "filter",
  mixed = F,
  id = NULL,
  weights = NULL,
  resid = F,
  corstr1 = "independence"
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{character outcome variable (string format)}

\item{value}{column name for value (string format)}

\item{confound}{character vector of confound variable (default is null)}

\item{family}{regression family (default is linear model)}

\item{type}{Whether to return only chemical effect (default is filter)}

\item{mixed}{Whether to run random intercept model}

\item{id}{id variable name}

\item{weights}{weights variable name}

\item{resid}{whether to output residuals}
}
\description{
Choose regression
}
\details{
\code{innerchem} Run regression model for chemical data

This is a function to run a single regression model for chemical data
}
\examples{
data(simchemdat)
# limit to one chemical
dat1 <- filter(simchemdat, chem == "chem1")
res <- innerchemlm(dat1, outcome = "out")
}
