% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemlm.R
\name{innerchemgee}
\alias{innerchemgee}
\title{innerchemgee}
\usage{
innerchemgee(
  data,
  outcome,
  id,
  weights,
  value = "value",
  confound = NULL,
  type = "filter",
  resid = F,
  corstr1 = "independence"
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{character outcome variable (string format)}

\item{id}{ID variable}

\item{value}{column name for value (string format)}

\item{confound}{character vector of confound variable (default is null)}

\item{type}{Whether to return only chemical effect (default is filter)}

\item{resid}{whether to output residuals}

\item{weight}{weight variable}
}
\description{
Running one regression model for chemical data
}
\details{
\code{innerchemlmer} Run GEE model (for MSM) with weights for chemical data

This is a function to run a single regression model for chemical data
}
\examples{
data(simchemdatmixed)
# limit to one chemical
dat1 <- dplyr::filter(simchemdatmixed, chem == "chem1")
res <- innerchemgee(dat1, outcome = "out", id = "id", weights = "weights")
}
