% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemlm.R
\name{innerchemlm}
\alias{innerchemlm}
\title{innerchemlm}
\usage{
innerchemlm(data, outcome, value = "value", confound = NULL,
  family = "gaussian")
}
\arguments{
\item{data}{dataset}

\item{outcome}{character outcome variable (string format)}

\item{value}{column name for value (string format)}

\item{confound}{character vector of confound variable (default is null)}

\item{family}{regression family (default is linear model)}
}
\description{
Running one regression model for chemical data
}
\details{
\code{innerchemlm} Run regression model for chemical data

This is a function to run a single regression model for chemical data
}
\examples{
data(simchemdat)
# limit to one chemical
dat1 <- filter(simchemdat, chem == "chem1")
res <- innerchemlm(dat1, outcome = "out")
}
