% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemlm.R
\name{innerchemlmer}
\alias{innerchemlmer}
\title{innerchemlmer}
\usage{
innerchemlmer(
  data,
  outcome,
  id,
  weights,
  value = "value",
  confound = NULL,
  type = "filter"
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{character outcome variable (string format)}

\item{id}{ID variable}

\item{value}{column name for value (string format)}

\item{confound}{character vector of confound variable (default is null)}

\item{type}{Whether to return only chemical effect (default is filter)}

\item{weight}{weight variable}
}
\description{
Running one regression model for chemical data
}
\details{
\code{innerchemlmer} Run linear mixed effects model with weights for chemical data

This is a function to run a single regression model for chemical data
}
\examples{
data(simchemdatmixed)
# limit to one chemical
dat1 <- dplyr::filter(simchemdatmixed, chem == "chem1")
res <- innerchemlmer(dat1, outcome = "out", id = "id", weights = "weights")
}
