% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simchemdat.R
\docType{data}
\name{simchemdat}
\alias{simchemdat}
\title{Chemical exposure test data}
\format{
A data frame with 200 rows and 4 variables
}
\description{
A dataset of 200 observations on 2 chemical exposures, 
one continuous outcome variable, and one continuous confounder
}
\details{
\itemize{
		\item outcome. outcome variable
		\item chem. chemical type
		\item value. value of each chemical
	\item confound. possible confounder
}
}
\keyword{datasets}
